Extract global files
